Wallrunning now implemented
Got wallrunning working. It does some stuff with gravity that might need tuned, but it functions pretty well. Again, checked everything but just the once. I also implemented the start of a HP regen system but then realized I was tired and didn't want to spend hours getting that working, so there's some variables in playController for that now. They don't do anything, so feel free to delete them and replace them with your own stuff if you need to.