Made a main menu!
Main Menu has been created. It has a play game button, a quit button, and a settings button. The settings button doesn't do anything. I mean, it doesn't exist yet, so that's to be expected.
I also made each of the three menus in the UI have buttons that link back to the main menu. Should all be working!